consolidated nvidia fd checking logic so both egl and direct backend can use it on their selected fds.
